Move noteburst interface to storage layer
Since it's an interface to an external resource, it belongs with
storage.
